Removed explicit carry computation
The explicit carry computation in the *getPost* function handled cases,
in which the output string was recurrent despite equaling a non
recurring number in the output base. Eg. 0.[1] equals 1 in the binary
system, but due to implementation it could happen, that 0.[1] would be
shown by toString. This is no longer the case with the new
implementation of the scale function.